chore(stdlib): split #1211 Blob/File code into fetch_blob.rs (file-size gate)
The #1211 additions pushed fetch.rs to 2124 lines, over the 2,000-line
lint cap. Move the Blob/File constructor + object-URL helpers into a
sibling `fetch_blob.rs` module; export `BLOB_REGISTRY`, `BlobData`,
`alloc_blob`, `handle_id`, `handle_to_f64`, `string_from_header`, and
`TAG_UNDEFINED` as `pub(crate)` so the new module can re-use the
existing handle ABI without reimplementing it.

fetch.rs now 1897 lines / fetch_blob.rs 250 lines. Buffer parity
suite still 94/94 PASS.
